+++ R/pip_grp_logic.R
@@ -24,10 +24,10 @@ pip_grp_logic <- function(country         = "ALL",
   reporting_level <- match.arg(reporting_level)
   group_by        <- match.arg(group_by)
 
-
   # Custom aggregations only supported at the national level
   # subgroups aggregations only supported for "all" countries
   country <- toupper(country)
+  year    <- toupper(year)
   if (group_by != "none") {
     reporting_level <- "all"
     if (!all(country %in% c("ALL", lkup$query_controls$region$values))) {
@@ -83,6 +83,12 @@ pip_grp_logic <- function(country         = "ALL",
       ppp             = NULL,
       lkup            = lkup
       )
+    # For now just rowbinding two dataframes, but we would need to use it more smartly in the future
+    fg_pip_master <- collapse::rowbind(fg_pip_master)
+
+    if (!data.table::is.data.table(fg_pip_master)) {
+      setDT(fg_pip_master)
+    }
 
     add_vars_out_of_pipeline(fg_pip_master, fill_gaps = TRUE, lkup = lkup)
 
@@ -290,7 +296,6 @@ pip_grp_helper <- function(lcv_country,
   if (nrow(out) == 0) {
     return(pipapi::empty_response_grp)
   }
-
   # Handles aggregated distributions
   if (reporting_level %in% c("national", "all")) {
     out <- add_agg_stats(out,
@@ -299,7 +304,9 @@ pip_grp_helper <- function(lcv_country,
 
   # Handle potential (insignificant) difference in poverty_line values that
   # may mess-up the grouping
-  out$poverty_line <- povline
+  # I don't think we need this out$poverty_line already has the correct values additionally,
+  # since povline is vectorized the below line does not work as expected
+  # out$poverty_line <- povline
 
   add_vars_out_of_pipeline(out, fill_gaps = TRUE, lkup = lkup)